projects/SynaFlow/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0

# API and cache management
cachetools==5.3.2
requests==2.31.0
httpx==0.25.1

# Machine learning and data science
numpy==1.24.3
scipy==1.10.1

# Utility libraries
tqdm==4.66.1
python-dateutil==2.8.2
asyncio==3.4.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Project-specific dependencies
synalinks>=0.1.0 